refactor : move function to lib
